Fix extra subtractions from recruitment scores
When there were previously invite links that had been used but have not
been used within the period of a particular leaderboard, they would
still be subtracted from the recruitment score. getDistinctInviteLinks
previously improved performance but no longer seems to be necessary.
The bug was within that function, but we can just remove it to fix the
issue.